+++ cmd/influx/bucket.go
@@ -12,8 +12,7 @@ import (
 type bucketSVCsFn func() (influxdb.BucketService, influxdb.OrganizationService, error)
 
 func cmdBucket(f *globalFlags, opt genericCLIOpts) *cobra.Command {
-	builder := newCmdBucketBuilder(newBucketSVCs, opt)
-	builder.globalFlags = f
+	builder := newCmdBucketBuilder(newBucketSVCs, f, opt)
 	return builder.cmd()
 }
 
@@ -32,15 +31,16 @@ type cmdBucketBuilder struct {
 	retention   string
 }
 
-func newCmdBucketBuilder(svcsFn bucketSVCsFn, opts genericCLIOpts) *cmdBucketBuilder {
+func newCmdBucketBuilder(svcsFn bucketSVCsFn, f *globalFlags, opts genericCLIOpts) *cmdBucketBuilder {
 	return &cmdBucketBuilder{
+		globalFlags:    f,
 		genericCLIOpts: opts,
 		svcFn:          svcsFn,
 	}
 }
 
 func (b *cmdBucketBuilder) cmd() *cobra.Command {
-	cmd := b.newCmd("bucket", nil, false)
+	cmd := b.newCmd("bucket", nil)
 	cmd.Short = "Bucket management commands"
 	cmd.TraverseChildren = true
 	cmd.Run = seeHelp
@@ -55,7 +55,7 @@ func (b *cmdBucketBuilder) cmd() *cobra.Command {
 }
 
 func (b *cmdBucketBuilder) cmdCreate() *cobra.Command {
-	cmd := b.newCmd("create", b.cmdCreateRunEFn, true)
+	cmd := b.newCmd("create", b.cmdCreateRunEFn)
 	cmd.Short = "Create bucket"
 
 	opts := flagOpts{
@@ -111,7 +111,7 @@ func (b *cmdBucketBuilder) cmdCreateRunEFn(*cobra.Command, []string) error {
 }
 
 func (b *cmdBucketBuilder) cmdDelete() *cobra.Command {
-	cmd := b.newCmd("delete", b.cmdDeleteRunEFn, true)
+	cmd := b.newCmd("delete", b.cmdDeleteRunEFn)
 	cmd.Short = "Delete bucket"
 
 	cmd.Flags().StringVarP(&b.id, "id", "i", "", "The bucket ID, required if name isn't provided")
@@ -166,7 +166,7 @@ func (b *cmdBucketBuilder) cmdDeleteRunEFn(cmd *cobra.Command, args []string) er
 }
 
 func (b *cmdBucketBuilder) cmdList() *cobra.Command {
-	cmd := b.newCmd("list", b.cmdListRunEFn, true)
+	cmd := b.newCmd("list", b.cmdListRunEFn)
 	cmd.Short = "List buckets"
 	cmd.Aliases = []string{"find", "ls"}
 
@@ -231,7 +231,7 @@ func (b *cmdBucketBuilder) cmdListRunEFn(cmd *cobra.Command, args []string) erro
 }
 
 func (b *cmdBucketBuilder) cmdUpdate() *cobra.Command {
-	cmd := b.newCmd("update", b.cmdUpdateRunEFn, true)
+	cmd := b.newCmd("update", b.cmdUpdateRunEFn)
 	cmd.Short = "Update bucket"
 
 	opts := flagOpts{
@@ -287,6 +287,12 @@ func (b *cmdBucketBuilder) cmdUpdateRunEFn(cmd *cobra.Command, args []string) er
 	}
 
 	return b.printBuckets(bucketPrintOpt{bucket: bkt})
+}
+
+func (b *cmdBucketBuilder) newCmd(use string, runE func(*cobra.Command, []string) error) *cobra.Command {
+	cmd := b.genericCLIOpts.newCmd(use, runE, true)
+	b.globalFlags.registerFlags(cmd)
+	return cmd
 }
 
 func (b *cmdBucketBuilder) registerPrintFlags(cmd *cobra.Command) {